refactor graph instructions to own file
