+++ client/src/components/Connections/components/Subgraph.js
@@ -2,14 +2,14 @@
 import React from 'react'
 import {compose} from 'redux'
 import {connect} from 'react-redux'
-import {withHandlers, withState} from 'recompose'
+import {withHandlers} from 'recompose'
 import GraphCompnent from 'react-graph-vis'
-import {Col, Row, Input} from 'reactstrap'
+import {Col, Row} from 'reactstrap'
 import {addNeighboursLimitSelector} from '../../../selectors'
 import Legend from '../../shared/Legend/Legend'
+import SubgraphInstructions from './SubgraphInstructions'
 import CompanyDetails from './../../shared/CompanyDetails'
 import {updateValue} from '../../../actions/sharedActions'
-import {setAddNeighboursLimit} from '../../../actions/connectionsActions'
 import SubgraphWrapper from '../dataWrappers/SubgraphWrapper'
 import {options as graphOptions, getNodeEid, addNeighbours, removeNodes} from './graph/utils'
 import {checkShaking, resetGesture} from './graph/gestures'
@@ -40,24 +40,17 @@ export type GraphEvent = {|
 export type OwnProps = {
   preloadNodes: boolean,
   limit: number,
-  limitSettingsOpen: boolean,
-  setLimitSettingsOpen: (open: boolean) => void,
 } & EntityProps &
   ConnectionProps
 
 type DispatchProps = {
   updateValue: Function,
-  setAddNeighboursLimit: Function,
 }
 type Handlers = {
   handleSelect: (e: GraphEvent) => void,
   handleDoubleClick: (e: GraphEvent) => void,
   handleDrag: (e: GraphEvent) => void,
   handleDragEnd: (e: GraphEvent) => void,
-  handleLimitChange: (e: Event) => void,
-  openLimitSettings: (e: Event) => void,
-  closeLimitSettings: (e: Event) => void,
-  submitOnEnter: (e: KeyboardEvent) => void,
 }
 
 type Props = OwnProps & SubgraphProps & DispatchProps & Handlers
@@ -77,40 +70,11 @@ const Subgraph = ({
   handleDoubleClick,
   handleDrag,
   handleDragEnd,
-  limit,
-  handleLimitChange,
-  limitSettingsOpen,
-  openLimitSettings,
-  closeLimitSettings,
-  submitOnEnter,
 }: Props) => (
   <div className="subgraph">
     <Row>
       <Col lg="7" md="12">
-        Ovládanie:
-        <ul>
-          <li>
-            <i>Ťahaním</i> mena ho premiestnite.
-          </li>
-          <li>
-            <i>Kliknutím</i> na meno zobrazíte detailné informácie v boxe pod grafom.
-          </li>
-          <li>
-            <i>Dvojkliknutím</i> na meno pridáte ďalších {limitSettingsOpen
-            ? <Input
-              className="limit-input"
-              value={limit || ''}
-              onChange={handleLimitChange}
-              onBlur={closeLimitSettings}
-              onKeyPress={submitOnEnter}
-              autoFocus
-            />
-            : <b className="limit" onClick={openLimitSettings}>{limit}</b>
-          } nezobrazených susedov.</li>
-          <li>
-            <i>Potrasením</i> mena ho odstránite zo schémy (spolu s jeho výlučnými susedmi).
-          </li>
-        </ul>
+        <SubgraphInstructions />
       </Col>
       <Col lg="5" md="12">
         <Legend />
@@ -142,9 +106,8 @@ export default compose(
     (state: State) => ({
       limit: addNeighboursLimitSelector(state),
     }),
-    {updateValue, setAddNeighboursLimit}
+    {updateValue}
   ),
-  withState('limitSettingsOpen', 'setLimitSettingsOpen', false),
   SubgraphWrapper,
   withHandlers({
     handleSelect: (props: OwnProps & DispatchProps & SubgraphProps) => ({nodes}: GraphEvent) => {
@@ -197,26 +160,5 @@ export default compose(
       }
     },
     handleDragEnd: () => resetGesture,
-    handleLimitChange: (props: OwnProps & DispatchProps & SubgraphProps) => (e: Event) => {
-      if (e.target instanceof HTMLInputElement) {
-        const {value} = e.target
-        const limit = Number(value)
-        if (value === '' || (Number.isInteger(limit) && limit >= 0)) {
-          props.setAddNeighboursLimit(value === '' ? null : limit)
-        }
-      }
-    },
-    openLimitSettings: (props: OwnProps & DispatchProps & SubgraphProps) => () =>
-      props.setLimitSettingsOpen(true),
-    closeLimitSettings: (props: OwnProps & DispatchProps & SubgraphProps) => () => {
-      props.limit || props.setAddNeighboursLimit(ADD_NEIGHBOURS_LIMIT)
-      props.setLimitSettingsOpen(false)
-    },
-    submitOnEnter: (props: OwnProps & DispatchProps & SubgraphProps) => (e: KeyboardEvent) => {
-      if (e.key === 'Enter') {
-        props.limit || props.setAddNeighboursLimit(ADD_NEIGHBOURS_LIMIT)
-        props.setLimitSettingsOpen(false)
-      }
-    },
   })
 )(Subgraph)
